requirements.txt
torch==1.10.0
pandas==1.3.4
sklearn==0.0
tqdm==4.62.3
wandb==0.12.7
transformers==4.12.5
logzero==1.7.0
optuna==2.10.0
click==8.0.3
ruamel.yaml==0.17.17
attrdict==2.0.1
